move storage keys to another file
