+++ src/app/core/components/Sidenav/Sidenav.tsx
@@ -18,10 +18,11 @@ import ReferralsWidget from 'app/referrals/components/ReferralsWidget/ReferralsW
 import { useAppSelector } from 'app/store/hooks';
 import InternxtLogo from 'assets/icons/big-logo.svg?react';
 import { t } from 'i18next';
-import localStorageService, { STORAGE_KEYS } from '../../../core/services/local-storage.service';
+import localStorageService from '../../../core/services/local-storage.service';
 import workspacesSelectors from '../../../store/slices/workspaces/workspaces.selectors';
 import SidenavItem from './SidenavItem/SidenavItem';
 import WorkspaceSelectorContainer from './WorkspaceSelectorContainer';
+import { STORAGE_KEYS } from '../../../core/services/storage-keys';
 
 export const HUNDRED_TB = 109951162777600;
