+++ src/app/core/factory/sdk/index.ts
@@ -6,15 +6,16 @@ import packageJson from '../../../../../package.json';
 import authService from '../../../auth/services/auth.service';
 import { AppDispatch } from '../../../store';
 import { userThunks } from '../../../store/slices/user';
-import { LocalStorageService, STORAGE_KEYS } from '../../services/local-storage.service';
+import LocalStorageService from '../../services/local-storage.service';
 import { Workspace } from '../../types';
 import { Checkout } from '@internxt/sdk/dist/payments';
 import { envConfig } from 'app/core/services/env.service';
+import { STORAGE_KEYS } from '../../services/storage-keys';
 
 export class SdkFactory {
   private static sdk: {
     dispatch: AppDispatch;
-    localStorage: LocalStorageService;
+    localStorage: typeof LocalStorageService;
     instance: SdkFactory;
     newApiInstance: SdkFactory;
   };
@@ -24,7 +25,7 @@ export class SdkFactory {
     this.apiUrl = apiUrl;
   }
 
-  public static initialize(dispatch: AppDispatch, localStorage: LocalStorageService): void {
+  public static initialize(dispatch: AppDispatch, localStorage: typeof LocalStorageService): void {
     this.sdk = {
       dispatch,
       localStorage,
