+++ src/app/drive/components/DriveExplorer/DriveExplorer.tsx
@@ -26,7 +26,8 @@ import BannerWrapper from '../../../banners/BannerWrapper';
 import deviceService from '../../../core/services/device.service';
 import envService from '../../../core/services/env.service';
 import errorService from '../../../core/services/error.service';
-import localStorageService, { STORAGE_KEYS } from '../../../core/services/local-storage.service';
+import localStorageService from '../../../core/services/local-storage.service';
+import { STORAGE_KEYS } from '../../../core/services/storage-keys';
 import navigationService from '../../../core/services/navigation.service';
 import RealtimeService, { SOCKET_EVENTS } from '../../../core/services/socket.service';
 import ClearTrashDialog from '../../../drive/components/ClearTrashDialog/ClearTrashDialog';
