+++ src/app/payment/utils/checkEnvironmentCode.ts
@@ -1,13 +1,15 @@
-import localStorageService, { STORAGE_KEYS } from 'app/core/services/local-storage.service';
+import localStorageService from 'app/core/services/local-storage.service';
+import { STORAGE_KEYS } from 'app/core/services/storage-keys';
 import paymentService from '../services/payment.service';
 import errorService from 'app/core/services/error.service';
 
 export const isEnvironmentThemeAvailable = async (onSuccess?: () => void): Promise<boolean> => {
-  const environmentInLocalStorage = localStorageService.get(STORAGE_KEYS.THEMES.ENVIRONMENT_THEME_AVAILABLE_LOCAL_STORAGE_KEY);
+  const environmentInLocalStorage = localStorageService.get(
+    STORAGE_KEYS.THEMES.ENVIRONMENT_THEME_AVAILABLE_LOCAL_STORAGE_KEY,
+  );
 
   if (environmentInLocalStorage === 'true') return true;
   try {
-
     const couponUsedResult = await paymentService.isCouponUsedByUser('PLANET85');
 
     if (couponUsedResult.couponUsed) {
