+++ src/app/payment/views/IntegratedCheckoutView/CheckoutViewWrapper.tsx
@@ -9,7 +9,8 @@ import { useCheckout } from 'hooks/checkout/useCheckout';
 import { useSignUp } from '../../../auth/components/SignUp/useSignUp';
 import envService, { envConfig } from '../../../core/services/env.service';
 import errorService from '../../../core/services/error.service';
-import localStorageService, { STORAGE_KEYS } from '../../../core/services/local-storage.service';
+import localStorageService from '../../../core/services/local-storage.service';
+import { STORAGE_KEYS } from '../../../core/services/storage-keys';
 import navigationService from '../../../core/services/navigation.service';
 import RealtimeService from '../../../core/services/socket.service';
 import AppError, { AppView, IFormValues } from '../../../core/types';
